Update to Omniverse Physics 110.0 and PhysX SDK 5.8.0

+++ physx/source/physxextensions/src/serialization/Xml/PsFastXml.h
@@ -22,7 +22,7 @@
 // (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 // OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 //
-// Copyright (c) 2008-2025 NVIDIA Corporation. All rights reserved.
+// Copyright (c) 2008-2026 NVIDIA Corporation. All rights reserved.
 // Copyright (c) 2004-2008 AGEIA Technologies, Inc. All rights reserved.
 // Copyright (c) 2001-2004 NovodeX AG. All rights reserved.
 
@@ -130,7 +130,7 @@ class FastXml
 			return true;
 		}
 
-		virtual void* allocate(uint32_t size)
+		virtual void* allocate(uint64_t size)
 		{
 			return PxGetBroadcastAllocator()->allocate(size, "FastXml", PX_FL);
 		}
